Improved UI/UX for droppr and receivr
