+++ client/src/components/index.js
@@ -1,6 +1,9 @@
-// Copyright (C) 2024 Droppr. All Rights Reserved.
+// Copyright (C) 2024 droppr. All Rights Reserved.
 //
 
 export * from './Dropper.jsx';
+export * from './FileContainer.jsx';
+export * from './FileItem.jsx';
 export * from './Footer.jsx';
 export * from './Header.jsx';
+export * from './Recipient.jsx';
